Changes to accommodate multiple users
The app now works with multiple users, although there isn't a way to add new users (yet). This involved adding user_id to the visited_airports table
in the database, and adjusting the relevant files listed below.

Also, it is possible to delete an airport without having to select it first.
More comments have been added, more to come!

 Changes to be committed:
	modified:   Model/Airport.cs
	modified:   Model/BusinessLogic.cs
	modified:   Model/Database.cs
	modified:   Model/IBusinessLogic.cs
	modified:   Model/IDatabase.cs
	modified:   UI/LoginPage.xaml
	modified:   UI/LoginPage.xaml.cs
	modified:   UI/MainPage.xaml.cs
	modified:   ViewModels/LoginViewModel.cs