Renamed MainPage.xaml as AirportsPage.xaml (finally!); changed its appearance to make it look more appealing (slightly)
On AirportsPage.xaml, added a converter so that rating now shows # of stars, not just a number. Also changed the
highlighting to a slightly more appealing shade. Adjusted column widths as well.

Beyond cosmetics, it is no longer necessary to select a row before clicking delete in the database.

Oh, did we mention this now works with mulitiple users? :-)

 Changes to be committed:
	modified:   Model/Database.cs
	new file:   SqlConsiderations.txt
	renamed:    UI/MainPage.xaml -> UI/AirportsPage.xaml
	renamed:    UI/MainPage.xaml.cs -> UI/AirportsPage.xaml.cs
	modified:   UI/MainTabbedPage.xaml
	new file:   UI/NumberToStarsConverter.cs
	new file:   UI/README.md
	modified:   ViewModels/LoginViewModel.cs