Changed navigation, resized icons
Instead of embedding LoginPage in a NavigationPage() and using
Navigation.PushAsync(new MainTabbedPage()), we now use
Application.Current.MainPage = new MainTabbedPage().

This solves a problem that was only visible on iOs, that the
tab bar contents were being duplicated.

The icons have resized to 30x30, so they don't look explosively
huge on iOS.