Added logout functionality, replaced Airports button text with icons
Added a +, pencil and logout icon to the Airports page to make it look better
It looks great on iOS now, but on Android the icons are showing up as black on
a purple background. Also added some fonts, but didn't actually use them as on
iOS they were showing up comically large.

 Changes to be committed:
	modified:   App.xaml
	modified:   Lab6-Starter.csproj
	modified:   MauiProgram.cs
	new file:   Resources/Fonts/Font Awesome 6 Free-Solid-900.otf
	new file:   Resources/Images/logout.png
	deleted:    Resources/Images/magnifyingglasssolid.svg
	new file:   Resources/Images/pencil.png
	new file:   Resources/Images/plus.png
	new file:   Resources/Styles/FancyIcons.xaml
	new file:   Resources/Styles/Icons.xaml
	modified:   UI/AirportsPage.xaml
	modified:   UI/AirportsPage.xaml.cs
	modified:   UI/LoginPage.xaml
	modified:   UI/Map.xaml